// src/app/core/api/model/cV.ts
/**
 * vacancy ANCFCC
 * API documentation
 *
 * The version of the OpenAPI document: v1
 * Contact: abdellah.elbekkali@digitransform.co
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Section } from './section';


export interface CV { 
    name?: string | null;
    sections?: Array<Section> | null;
}